nixos: switch pinning method
Setting nixPath has bad UX, since it only changes the default the
NIX_PATH environment variable is set to at login. That means any changes
require a relogin to take effect. See

https://logs.nix.samueldr.com/nixos/2019-11-16#2791484

With this new way we can keep nixPath constant and instead just change
the symlinked nixpkgs. This is still pure. Only downside is that the
first time switching to this method will still require a re-login to set
the nixPath to the new symlink. That is why I also added the -I option
to `home-manager switch` in the rebuild script.