neovim: do not fetch hunspell
Unfortunately the website[1] and the corresponding github repo were
taken down. There is still an (rather old) archive[2] of it.

Luckily it seems that since then the libreoffice dictionaries have been
incorporated[3] into upstream vim in as a result of [4]. Therefore,
hunspell is no longer necessary. The annoying thing is that this will
break all my prior configurations, whenever I want to roll back I'll
have to comment out the hunspell part now.

[1] https://1042.ch/spell/hun-de-DE.utf-8.spl
[2] https://web.archive.org/web/*/https://github.com/ganwell/vim-hunspell-dicts
[3] https://github.com/vim/vim/commit/2a953fcf107d24229fec8af820ee62c908caafbf
[4] https://github.com/vim/vim/issues/4314#issuecomment-515936753