Use the old nix CLI in the rebuild script
Using the new CLI makes it a pain to switch between old and new
configurations since nix is currently taken from the environment. An old
configuration would have an older version of nix and couldn't run the
rebuild script.

Using the old CLI avoid this for now since newer nix versions still
support it.

It would be nicer to take the nix version from the pinned nixpkgs, but
that poses bootstrapping problems.